requirements.txt
psycopg2-binary==2.9.9
httpx==0.25.2
openai==1.3.10
python-dotenv==1.0.0
